\name{denoiseTDA-package}
\alias{denoiseTDA-package}
\alias{denoiseTDA}
\docType{package}
\title{
<<<<<<< HEAD
   Denoising Data and some functions for Topological Data Analysis
 }
 \description{
   This package contains two algorithms, Kloke's algorithm^(1) and mean shift algorithm^(2), which de-noise data for further topological data analysis. 
 }
 \details{
   KlokeKDE, KlokeAlg, and MeanShiftAlg are main functions for denoising, the rest are supporting function to test the results. 
 }
 \author{
 Uriya

 Maintainer: Your Name <your@email.com>
 }
 \references{
    (1) Kloke, J. N., Carlsson, G., Kerckhoff, S., Mazzeo, R. (2010). Methods and applications of topological data analysis. Stanford University.\\
    (2) Fukunaga, K., & Hostetler, L.D. (1975). The estimation of the gradient of a density function, with applications in pattern recognition. IEEE Trans. Inf. Theory, 21, 32-40.
 }
 \keyword{ package }
 \seealso{
   Optional links to other man pages
 }
 \examples{
   \dontrun{
      ## Optional simple examples of the most important functions
      ## These can be in \dontrun{} and \donttest{} blocks.   
   }

 }
=======
  Denoising Data and some functions for 
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
>>>>>>> parent of c6a01eb... Update denoiseTDA-package.Rd
