\name{denoiseTDA-package}
\alias{denoiseTDA-package}
\alias{denoiseTDA}
\docType{package}
\title{
  Denoising Data and some functions for Topological Data Analysis
}
\description{
  This package contains two algorithms, Klokes algorithm (1) and mean shift algorithm (2), which de-noise data for further topological data analysis.
}
\details{
  KlokeKDE, KlokeAlg, and MeanShiftAlg are main functions for denoising, the rest are supporting function to test the results. 
}

\references{
   (1) Kloke, J. N., Carlsson, G., Kerckhoff, S., Mazzeo, R. (2010). Methods and applications of topological data analysis. Stanford University.
   
   (2) Fukunaga, K., & Hostetler, L.D. (1975). The estimation of the gradient of a density function, with applications in pattern recognition. IEEE Trans. Inf. Theory, 21, 32-40.
}
\keyword{ package }

 \seealso{
   Optional links to other man pages
 }
 \examples{
   \dontrun{
      ## Optional simple examples of the most important functions
      ## These can be in \dontrun{} and \donttest{} blocks.   
   }

 }

